extract getTxData to utils
